+++ packages/moonstone/ExpandableItem/ExpandableSpotlightDecorator.js
@@ -22,7 +22,7 @@ const defaultConfig = {
 	 * @memberof moonstone/ExpandableItem.ExpandableSpotlightDecorator.defaultConfig
 	 * @private
 	 */
-	getSelectedNode: null,
+	getChildFocusTarget: null,
 
 	/**
 	 * When `true` and used in conjunction with `noAutoFocus` when `false`, the contents of the
@@ -45,7 +45,7 @@ const defaultConfig = {
  * @private
  */
 const ExpandableSpotlightDecorator = hoc(defaultConfig, (config, Wrapped) => {
-	const {getSelectedNode, noPointerMode} = config;
+	const {getChildFocusTarget, noPointerMode} = config;
 
 	return class extends React.Component {
 		static displayName = 'ExpandableSpotlightDecorator'
@@ -79,11 +79,13 @@ const ExpandableSpotlightDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			const current = Spotlight.getCurrent();
 			if (this.containerNode.contains(current) || document.activeElement === document.body) {
 				const contents = this.containerNode.querySelector('[data-expandable-container]');
-				const selectedNode = getSelectedNode ? getSelectedNode(contents, this.props) : null;
+				const selectedNode = getChildFocusTarget ? getChildFocusTarget(contents, this.props) : null;
 				if (contents && !this.props.noAutoFocus && !contents.contains(current)) {
+					let fucused = false;
 					if (selectedNode) {
-						Spotlight.focus(selectedNode);
-					} else {
+						fucused = Spotlight.focus(selectedNode);
+					}
+					if (!fucused) {
 						Spotlight.focus(contents.dataset.containerId);
 					}
 				}