+++ packages/moonstone/CHANGELOG.md
@@ -12,9 +12,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 - `moonstone/Marquee.Marquee` to be `moonstone/Marquee.MarqueeBase`
 - `moonstone/VirtualGridList.GridListImageItem` in favor of `moonstone/GridListImageItem`
-- `moonstone/ExpandableSpotlightDecorator` to add `getSelectedNode` variable from config to focus the selected index as the default focus
-- `moonstone/ExpandableList` to add HOC prameter for finding selcted node
-- `moonstone/Expandable` to add HOC prop. If `getSelectedNode` is defined in config, it will be used for finding selected node in `ExpandableSpotlightDecorator`
+- `moonstone/ExpandableList` to allow restoring focus to the first selected item when opening
 
 ### Fixed